move most of fs ops to cli/js/ops/fs/
