+++ cli/js/deno.ts
@@ -9,18 +9,18 @@ export {
   writeAllSync
 } from "./buffer.ts";
 export { build, OperatingSystem, Arch } from "./build.ts";
-export { chmodSync, chmod } from "./chmod.ts";
-export { chownSync, chown } from "./chown.ts";
+export { chmodSync, chmod } from "./ops/fs/chmod.ts";
+export { chownSync, chown } from "./ops/fs/chown.ts";
 export { transpileOnly, compile, bundle } from "./compiler_api.ts";
 export { inspect } from "./console.ts";
-export { copyFileSync, copyFile } from "./copy_file.ts";
+export { copyFileSync, copyFile } from "./ops/fs/copy_file.ts";
 export {
   Diagnostic,
   DiagnosticCategory,
   DiagnosticItem,
   DiagnosticMessageChain
 } from "./diagnostics.ts";
-export { chdir, cwd } from "./dir.ts";
+export { chdir, cwd } from "./ops/fs/dir.ts";
 export { applySourceMap, formatDiagnostics } from "./ops/errors.ts";
 export { errors } from "./errors.ts";
 export { FileInfo } from "./file_info.ts";
@@ -59,16 +59,16 @@ export {
   ReadWriteCloser,
   ReadWriteSeeker
 } from "./io.ts";
-export { linkSync, link } from "./link.ts";
+export { linkSync, link } from "./ops/fs/link.ts";
 export {
   makeTempDirSync,
   makeTempDir,
   makeTempFileSync,
   makeTempFile,
   MakeTempOptions
-} from "./make_temp.ts";
+} from "./ops/fs/make_temp.ts";
 export { metrics, Metrics } from "./ops/runtime.ts";
-export { mkdirSync, mkdir, MkdirOptions } from "./mkdir.ts";
+export { mkdirSync, mkdir, MkdirOptions } from "./ops/fs/mkdir.ts";
 export {
   Addr,
   connect,
@@ -100,20 +100,20 @@ export {
 export { openPlugin } from "./plugins.ts";
 export { kill } from "./ops/process.ts";
 export { run, RunOptions, Process, ProcessStatus, Signal } from "./process.ts";
-export { readdirSync, readdir } from "./read_dir.ts";
+export { readdirSync, readdir } from "./ops/fs/read_dir.ts";
 export { readFileSync, readFile } from "./read_file.ts";
-export { readlinkSync, readlink } from "./read_link.ts";
-export { realpathSync, realpath } from "./realpath.ts";
-export { removeSync, remove, RemoveOptions } from "./remove.ts";
-export { renameSync, rename } from "./rename.ts";
+export { readlinkSync, readlink } from "./ops/fs/read_link.ts";
+export { realpathSync, realpath } from "./ops/fs/realpath.ts";
+export { removeSync, remove, RemoveOptions } from "./ops/fs/remove.ts";
+export { renameSync, rename } from "./ops/fs/rename.ts";
 export { resources, close } from "./ops/resources.ts";
 export { signal, signals, SignalStream } from "./signals.ts";
-export { statSync, lstatSync, stat, lstat } from "./stat.ts";
-export { symlinkSync, symlink } from "./symlink.ts";
+export { statSync, lstatSync, stat, lstat } from "./ops/fs/stat.ts";
+export { symlinkSync, symlink } from "./ops/fs/symlink.ts";
 export { connectTLS, listenTLS } from "./tls.ts";
-export { truncateSync, truncate } from "./truncate.ts";
+export { truncateSync, truncate } from "./ops/fs/truncate.ts";
 export { isatty, setRaw } from "./ops/tty.ts";
-export { utimeSync, utime } from "./utime.ts";
+export { utimeSync, utime } from "./ops/fs/utime.ts";
 export { version } from "./version.ts";
 export { writeFileSync, writeFile, WriteFileOptions } from "./write_file.ts";
 export const args: string[] = [];
