+++ cli/js/ops/fs/chown.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { sendSync, sendAsync } from "./ops/dispatch_json.ts";
+import { sendSync, sendAsync } from "../dispatch_json.ts";
 
 /** Synchronously change owner of a regular file or directory. Linux/Mac OS
  * only at the moment.
