+++ cli/js/ops/fs/make_temp.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { sendSync, sendAsync } from "./ops/dispatch_json.ts";
+import { sendSync, sendAsync } from "../dispatch_json.ts";
 
 export interface MakeTempOptions {
   /** Directory where the temporary directory should be created (defaults to
