+++ cli/js/ops/fs/remove.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { sendSync, sendAsync } from "./ops/dispatch_json.ts";
+import { sendSync, sendAsync } from "../dispatch_json.ts";
 
 export interface RemoveOptions {
   /** Defaults to `false`. If set to `true`, path will be removed even if
